zakrzewski_spark_task_project/requirements.txt
pyspark==3.4.1
pyyaml==6.0.1
requests==2.32.3
pytest==7.4.0
chispa==0.9.2